refactoring of node and transport packets for improving code readability
