+++ src/mesh/domain/config.go
@@ -3,14 +3,8 @@ package domain
 import (
 	"time"
 
-	"github.com/satori/go.uuid"
 	"github.com/skycoin/skycoin/src/cipher"
 )
-
-type RouteConfig struct {
-	ID    uuid.UUID
-	Peers []cipher.PubKey
-}
 
 type NodeConfig struct {
 	PubKey                        cipher.PubKey
@@ -21,8 +15,4 @@ type NodeConfig struct {
 	TimeToAssembleMessage         time.Duration
 	TransportMessageChannelLength int
 	//ChaCha20Key                   [32]byte
-}
-
-type TransportConfig struct {
-	SendChannelLength uint32
 }
