--- src/mesh/domain/peer.go
@@ -1,8 +1,0 @@
-package domain
-
-import "github.com/skycoin/skycoin/src/cipher"
-
-type Peer struct {
-	Peer cipher.PubKey
-	Info string
-}
