+++ src/mesh/serialize/serialize.go
@@ -5,9 +5,7 @@ import (
 	"fmt"
 	"log"
 	"reflect"
-)
 
-import (
 	"github.com/skycoin/skycoin/src/cipher/encoder"
 )
 
@@ -17,99 +15,97 @@ const messagePrefixLength = 1
 type MessagePrefix [messagePrefixLength]byte
 
 type Serializer struct {
-	messageIdMap        map[reflect.Type]MessagePrefix
-	messageIdReverseMap map[MessagePrefix]reflect.Type
+	messageTypeToPrefix map[reflect.Type]MessagePrefix
+	messagePrefixToType map[MessagePrefix]reflect.Type
 }
 
 func NewSerializer() *Serializer {
-	ret := &Serializer{}
-	ret.messageIdMap = make(map[reflect.Type]MessagePrefix)
-	ret.messageIdReverseMap = make(map[MessagePrefix]reflect.Type)
-	return ret
+	serializer := &Serializer{}
+	serializer.messageTypeToPrefix = make(map[reflect.Type]MessagePrefix)
+	serializer.messagePrefixToType = make(map[MessagePrefix]reflect.Type)
+	return serializer
 }
 
 // Register a message struct for recognition by the message handlers.
-func (self *Serializer) RegisterMessageForSerialization(prefix MessagePrefix, msg interface{}) {
-	t := reflect.TypeOf(msg)
-	id := MessagePrefix{}
-	copy(id[:], prefix[:])
-	_, exists := self.messageIdReverseMap[id]
+func (self *Serializer) RegisterMessageForSerialization(messagePrefix MessagePrefix, message interface{}) {
+	messageType := reflect.TypeOf(message)
+	messagePrefixBuf := MessagePrefix{}
+	copy(messagePrefixBuf[:], messagePrefix[:])
+	_, exists := self.messagePrefixToType[messagePrefixBuf]
 	if exists {
 		log.Panicf("Attempted to register message prefix %s twice",
-			string(id[:]))
+			string(messagePrefixBuf[:]))
 	}
-	_, exists = self.messageIdMap[t]
+	_, exists = self.messageTypeToPrefix[messageType]
 	if exists {
-		log.Panicf("Attempts to register message type %v twice", t)
+		log.Panicf("Attempts to register message type %v twice", messageType)
 	}
-	self.messageIdMap[t] = id
-	self.messageIdReverseMap[id] = t
+	self.messageTypeToPrefix[messageType] = messagePrefixBuf
+	self.messagePrefixToType[messagePrefixBuf] = messageType
 }
 
-func (self *Serializer) UnserializeMessage(msg []byte) (interface{}, error) {
-	msgId := [1]byte{}
-	if len(msg) < len(msgId) {
-		return nil, errors.New("Not enough data to read msg id")
+func (self *Serializer) UnserializeMessage(messageData []byte) (interface{}, error) {
+	messagePrefix := [1]byte{}
+	if len(messageData) < len(messagePrefix) {
+		return nil, errors.New("Not enough data to read message prefix")
 	}
-	copy(msgId[:], msg[:len(msgId)])
-	msg = msg[len(msgId):]
+	copy(messagePrefix[:], messageData[:len(messagePrefix)])
+	messageData = messageData[len(messagePrefix):]
 
-	t, succ := self.messageIdReverseMap[msgId]
-	if !succ {
-		return nil, fmt.Errorf("Unknown message %s received", string(msgId[:]))
+	messageType, exists := self.messagePrefixToType[messagePrefix]
+	if !exists {
+		return nil, fmt.Errorf("Unknown message %s received", string(messagePrefix[:]))
 	}
-	var m interface{}
-	var v reflect.Value = reflect.New(t)
+	var reflectValue reflect.Value = reflect.New(messageType)
 	//logger.Debug("Giving %d bytes to the decoder", len(msg))
-	used, err := self.deserializeMessage(msg, v)
+	usedBytesCount, err := self.deserializeMessage(messageData, reflectValue)
 	if err != nil {
 		return nil, err
 	}
-	if used != len(msg) {
+	if usedBytesCount != len(messageData) {
 		return nil, errors.New("Data buffer was not completely decoded")
 	}
-	m, succ = (v.Elem().Interface()).(interface{})
-	if !succ {
+	message, exists := (reflectValue.Elem().Interface()).(interface{})
+	if !exists {
 		// This occurs only when the user registers an interface that does
 		// match the interface{} interface.  They should have known about this
 		// earlier via a call to VerifyMessages
 		log.Panic("interface{} obtained from map does not match interface{} interface")
-		return nil, errors.New("MessageIdMaps contain non-interface{}")
+		return nil, errors.New("messagePrefixToType contain non-interface{}")
 	}
-	return m, nil
+	return message, nil
 }
 
 // Wraps encoder.DeserializeRawToValue and traps panics as an error
-func (self *Serializer) deserializeMessage(msg []byte, v reflect.Value) (n int, e error) {
+func (self *Serializer) deserializeMessage(messageData []byte, reflectValue reflect.Value) (usedBytesCount int, err error) {
 	defer func() {
 		if r := recover(); r != nil {
 			switch x := r.(type) {
 			case string:
-				e = errors.New(x)
+				err = errors.New(x)
 			case error:
-				e = x
+				err = x
 			default:
-				e = errors.New("interface{} deserialization failed")
+				err = errors.New("interface{} deserialization failed")
 			}
 		}
 	}()
-	n, e = encoder.DeserializeRawToValue(msg, v)
+	usedBytesCount, err = encoder.DeserializeRawToValue(messageData, reflectValue)
 	return
 }
 
-// Packgs a interface{} into []byte containing length, id and data
-func (self *Serializer) SerializeMessage(msg interface{}) []byte {
-	t := reflect.TypeOf(msg)
-	msgId, succ := self.messageIdMap[t]
-	if !succ {
-		txt := "Attempted to serialize message type not in MessageIdMap: %v"
-		log.Panicf(txt, t)
+// Packs a interface{} into []byte containing length, id and data
+func (self *Serializer) SerializeMessage(message interface{}) []byte {
+	messageType := reflect.TypeOf(message)
+	messagePrefix, exists := self.messageTypeToPrefix[messageType]
+	if !exists {
+		log.Panicf("Attempted to serialize message with unknown type: %v", messageType)
 	}
-	bMsg := encoder.Serialize(msg)
+	serializedMessageData := encoder.Serialize(message)
 
 	// message length
-	m := make([]byte, 0)
-	m = append(m, msgId[:]...) // message id
-	m = append(m, bMsg...)     // message bytes
-	return m
+	finalMessage := make([]byte, 0)
+	finalMessage = append(finalMessage, messagePrefix[:]...)      // message prefix
+	finalMessage = append(finalMessage, serializedMessageData...) // message bytes
+	return finalMessage
 }
